[REFACT] Period컴포넌트 분리 #NADA-52
